break tests into two files and minor cleanup
